pyproject.toml
[project]
name = "capitok"
version = "0.1.0"
description = "Capitok memory gateway MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic-settings>=2.6.0",
  "psycopg[binary]>=3.2.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
